added cooldown for gate keeper attacks

+++ Source/Logic/GateKeeperSource/src/GateKeeper.h
@@ -42,14 +42,14 @@ class GateKeeper : public Marx::VEntity, public Creature
 		virtual void setRange(int r);
 		virtual void setHealth(int h);
 		virtual void setAttack(int a);
-		virtual void setAttackSpeed(int as);
+		virtual void setAttackSpeed(float as);
 		virtual void setXSpeed(float x);
 		virtual void setYSpeed(float y);
         virtual void setSpeed(int _speed);
 		virtual int getRange();
 		virtual int getHealth();
 		virtual int getAttack();
-		virtual int getAttackSpeed();
+		virtual float getAttackSpeed();
 		virtual int getMovementSpeed();
         virtual int getSpeed();
 		virtual void turn();
@@ -73,7 +73,7 @@ class GateKeeper : public Marx::VEntity, public Creature
 		int _type;
 		int _health;
 		int _attack;
-		int _attackSpeed;
+		float _attackSpeed;
 		int _movementSpeed;
 		float _xSpeed;
 		float _ySpeed;
@@ -91,6 +91,7 @@ class GateKeeper : public Marx::VEntity, public Creature
 		BlockZone steppedTile;
 		bool soundActive;
 		Animation *gkAnimation;
+
 
 
 };
